Refactor to be less mono-file
This started out as a quick hack in a single file, but it's grown a
little so this is a step to add a little more structure for sanity's
sake and testability.
