+++ src/dataform.js
@@ -1,4 +1,12 @@
-// Parse a dataform partition by clause to get a dbt version
+import path from 'path'
+import dataform from '@dataform/core'
+
+import { getFiles } from './fs.js'
+import { saferEval } from './utils.js'
+
+/**
+ * Parse a dataform partition by clause to get a dbt version
+ */
 export const parsePartitionBy = (value) => {
   if (!value) return undefined
   // e.g. TIMESTAMP_TRUNC(<date_column>, MONTH)
@@ -38,4 +46,46 @@ export const parsePartitionBy = (value) => {
   }
 
   throw new Error(`Unable to parse partitioning clause: ${value}`)
+}
+
+/**
+ * Resolve dataform includes for use in parsing
+ */
+export const resolveIncludes = async (root) => {
+  const files = await getFiles(path.resolve(root, 'includes'), false)
+  return files.reduce(async (getAcc, file) => {
+    const acc = await getAcc
+    acc[file.file.base] = await import(file.file.absolute)
+    return acc
+  }, Promise.resolve({}))
+}
+
+/**
+ * Extract parsed dataform data.
+ */
+export const parseExtractor = (includes) => (content, fileName) => {
+  const parsed = dataform.compiler(content, fileName)
+
+  // Collect payloads pushed to dataform.sqlxAction
+  const collected = []
+  const collector = {
+    sqlxAction: (payload) => collected.push(payload),
+  }
+
+  // Execute eval-ed code, passing in required localised globals
+  saferEval(`
+    function unpack(includes, dataform) {
+      const { ${Object.keys(includes).join(', ')} } = includes;
+      ${parsed}
+    }
+  `)(includes, collector)
+
+  // Sanity check - should only be called once
+  if (collected.length !== 1) {
+    throw new Error(
+      `dataform parse extraction failed, ${collected.length} calls`,
+    )
+  }
+
+  return collected[0]
 }
